+++ packages/crypto/src/keys/rsa/der.ts
@@ -13,8 +13,11 @@ interface Decoder {
 }
 
 const decoders: Record<number, Decoder> = {
+  0x0: readSequence,
+  0x1: readSequence,
   0x2: readInteger,
   0x3: readBitString,
+  0x4: readOctetString,
   0x5: readNull,
   0x6: readObjectIdentifier,
   0x10: readSequence,
@@ -96,13 +99,53 @@ function readInteger (buf: Uint8Array, context: Context): Uint8Array {
   return Uint8Array.from(vals)
 }
 
-function readObjectIdentifier (buf: Uint8Array, context: Context): string[] {
+function readObjectIdentifier (buf: Uint8Array, context: Context): string {
   const count = readLength(buf, context)
+  const finalOffset = context.offset + count
 
-  // skip OID
-  context.offset += count
+  const byte = buf[context.offset]
+  context.offset++
 
-  return ['oid-unimplemented']
+  let val1 = 0
+  let val2 = 0
+
+  if (byte < 40) {
+    val1 = 0
+    val2 = byte
+  } else if (byte < 80) {
+    val1 = 1
+    val2 = byte - 40
+  } else {
+    val1 = 2
+    val2 = byte - 80
+  }
+
+  let oid = `${val1}.${val2}`
+  let num: number[] = []
+
+  while (context.offset < finalOffset) {
+    const byte = buf[context.offset]
+    context.offset++
+
+    // remove msb
+    num.push(byte & 0b01111111)
+
+    if (byte < 128) {
+      num.reverse()
+
+      // reached the end of the encoding
+      let val = 0
+
+      for (let i = 0; i < num.length; i++) {
+        val += num[i] << (i * 7)
+      }
+
+      oid += `.${val}`
+      num = []
+    }
+  }
+
+  return oid
 }
 
 function readNull (buf: Uint8Array, context: Context): null {
@@ -115,7 +158,7 @@ function readBitString (buf: Uint8Array, context: Context): any {
   const length = readLength(buf, context)
   const unusedBits = buf[context.offset]
   context.offset++
-  const bytes = buf.subarray(context.offset, context.offset + length)
+  const bytes = buf.subarray(context.offset, context.offset + length - 1)
   context.offset += length
 
   if (unusedBits !== 0) {
@@ -123,9 +166,15 @@ function readBitString (buf: Uint8Array, context: Context): any {
     throw new Error('Unused bits in bit string is unimplemented')
   }
 
-  return decodeDer(bytes, {
-    offset: 0
-  })
+  return bytes
+}
+
+function readOctetString (buf: Uint8Array, context: Context): any {
+  const length = readLength(buf, context)
+  const bytes = buf.subarray(context.offset, context.offset + length)
+  context.offset += length
+
+  return bytes
 }
 
 function encodeNumber (value: number): Uint8ArrayList {
@@ -163,7 +212,7 @@ function encodeLength (bytes: { byteLength: number }): Uint8Array | Uint8ArrayLi
 export function encodeInteger (value: Uint8Array | Uint8ArrayList): Uint8ArrayList {
   const contents = new Uint8ArrayList()
 
-  const mask = parseInt('10000000', 2)
+  const mask = 0b10000000
   const positive = (value.subarray()[0] & mask) === mask
 
   if (positive) {
@@ -195,7 +244,15 @@ export function encodeBitString (value: Uint8Array | Uint8ArrayList): Uint8Array
   )
 }
 
-export function encodeSequence (values: Array<Uint8Array | Uint8ArrayList>): Uint8ArrayList {
+export function encodeOctetString (value: Uint8Array | Uint8ArrayList): Uint8ArrayList {
+  return new Uint8ArrayList(
+    Uint8Array.from([0x04]),
+    encodeLength(value),
+    value
+  )
+}
+
+export function encodeSequence (values: Array<Uint8Array | Uint8ArrayList>, tag = 0x30): Uint8ArrayList {
   const output = new Uint8ArrayList()
 
   for (const buf of values) {
@@ -205,7 +262,7 @@ export function encodeSequence (values: Array<Uint8Array | Uint8ArrayList>): Uin
   }
 
   return new Uint8ArrayList(
-    Uint8Array.from([0x30]),
+    Uint8Array.from([tag]),
     encodeLength(output),
     output
   )