+++ packages/integration-tests/test/dht.node.ts
@@ -1,8 +1,10 @@
 /* eslint-env mocha */
 
 import { yamux } from '@chainsafe/libp2p-yamux'
+import { identify, type Identify } from '@libp2p/identify'
 import { kadDHT, passthroughMapper } from '@libp2p/kad-dht'
 import { mplex } from '@libp2p/mplex'
+import { ping } from '@libp2p/ping'
 import { plaintext } from '@libp2p/plaintext'
 import { tcp } from '@libp2p/tcp'
 import { multiaddr } from '@multiformats/multiaddr'
@@ -13,6 +15,7 @@ import pWaitFor from 'p-wait-for'
 import { fromString as uint8ArrayFromString } from 'uint8arrays/from-string'
 import type { Libp2p, PeerId } from '@libp2p/interface'
 import type { KadDHT } from '@libp2p/kad-dht'
+import type { PingService } from '@libp2p/ping'
 import type { Multiaddr } from '@multiformats/multiaddr'
 import type { Libp2pOptions } from 'libp2p'
 
@@ -62,7 +65,9 @@ describe('DHT subsystem operates correctly', () => {
             protocol: subsystemMulticodecs[0],
             peerInfoMapper: passthroughMapper,
             allowQueryWithZeroPeers: true
-          })
+          }),
+          ping: ping(),
+          identify: identify()
         }
       })
 
@@ -85,7 +90,9 @@ describe('DHT subsystem operates correctly', () => {
             protocol: subsystemMulticodecs[0],
             peerInfoMapper: passthroughMapper,
             allowQueryWithZeroPeers: true
-          })
+          }),
+          ping: ping(),
+          identify: identify()
         }
       })
 
@@ -145,7 +152,7 @@ describe('DHT subsystem operates correctly', () => {
   it('kad-dht should discover other peers', async () => {
     const deferred = pDefer()
 
-    const getConfig = (): Libp2pOptions<{ dht: KadDHT }> => ({
+    const getConfig = (): Libp2pOptions<{ dht: KadDHT, ping: PingService, identify: Identify }> => ({
       addresses: {
         listen: [
           listenAddr.toString()
@@ -156,7 +163,9 @@ describe('DHT subsystem operates correctly', () => {
           protocol: subsystemMulticodecs[0],
           peerInfoMapper: passthroughMapper,
           allowQueryWithZeroPeers: true
-        })
+        }),
+        ping: ping(),
+        identify: identify()
       }
     })