+++ packages/metrics-prometheus/CHANGELOG.md
@@ -81,6 +81,29 @@
   * devDependencies
     * @libp2p/interface-compliance-tests bumped from ^5.3.0 to ^5.3.1
 
+## [4.3.16](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.15...prometheus-metrics-v4.3.16) (2025-04-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.7.0 to ^2.8.0
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.0.15 to ^5.1.0
+    * @libp2p/interface-compliance-tests bumped from ^6.4.2 to ^6.4.3
+    * @libp2p/logger bumped from ^5.1.13 to ^5.1.14
+    * @libp2p/peer-id bumped from ^5.1.0 to ^5.1.1
+
+## [4.3.15](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.14...prometheus-metrics-v4.3.15) (2025-03-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.1 to ^6.4.2
+
 ## [4.3.14](https://github.com/libp2p/js-libp2p/compare/prometheus-metrics-v4.3.13...prometheus-metrics-v4.3.14) (2025-03-12)