+++ packages/pnet/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Changelog
 
+## [2.0.33](https://github.com/libp2p/js-libp2p/compare/pnet-v2.0.32...pnet-v2.0.33) (2025-04-09)
+
+
+### Bug Fixes
+
+* update stream deps ([#3055](https://github.com/libp2p/js-libp2p/issues/3055)) ([b2124c2](https://github.com/libp2p/js-libp2p/commit/b2124c2db02d7870b958f294da42ec79084818a3))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/crypto bumped from ^5.0.15 to ^5.1.0
+    * @libp2p/interface bumped from ^2.7.0 to ^2.8.0
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.2 to ^6.4.3
+    * @libp2p/logger bumped from ^5.1.13 to ^5.1.14
+    * @libp2p/peer-id bumped from ^5.1.0 to ^5.1.1
+
+## [2.0.32](https://github.com/libp2p/js-libp2p/compare/pnet-v2.0.31...pnet-v2.0.32) (2025-03-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @libp2p/interface-compliance-tests bumped from ^6.4.1 to ^6.4.2
+
 ## [2.0.31](https://github.com/libp2p/js-libp2p/compare/pnet-v2.0.30...pnet-v2.0.31) (2025-03-12)