+++ packages/upnp-nat/CHANGELOG.md
@@ -64,6 +64,31 @@
   * dependencies
     * @libp2p/utils bumped from ^5.2.4 to ^5.2.5
 
+## [3.1.12](https://github.com/libp2p/js-libp2p/compare/upnp-nat-v3.1.11...upnp-nat-v3.1.12) (2025-04-09)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface bumped from ^2.7.0 to ^2.8.0
+    * @libp2p/interface-internal bumped from ^2.3.9 to ^2.3.10
+    * @libp2p/utils bumped from ^6.6.0 to ^6.6.1
+  * devDependencies
+    * @libp2p/crypto bumped from ^5.0.15 to ^5.1.0
+    * @libp2p/logger bumped from ^5.1.13 to ^5.1.14
+    * @libp2p/peer-id bumped from ^5.1.0 to ^5.1.1
+
+## [3.1.11](https://github.com/libp2p/js-libp2p/compare/upnp-nat-v3.1.10...upnp-nat-v3.1.11) (2025-03-18)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @libp2p/interface-internal bumped from ^2.3.8 to ^2.3.9
+    * @libp2p/utils bumped from ^6.5.8 to ^6.6.0
+
 ## [3.1.10](https://github.com/libp2p/js-libp2p/compare/upnp-nat-v3.1.9...upnp-nat-v3.1.10) (2025-03-12)